tests/wpt/web-platform-tests/css/cssom-view/matchMedia-display-none-iframe.html

<!doctype html>
<title>CSS Test: matchMedia works on display: none iframes</title>
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-window-matchmedia">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<link rel="author" title="Mozilla" href="https://mozilla.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  function frameLoaded(frame) {
    test(function() {
      assert_true(frame.contentWindow.matchMedia("all").matches);
    }, "matchMedia should work in display: none iframes");
    test(function() {
      assert_true(frame.contentWindow.matchMedia("(min-width: 0)").matches);
      assert_true(!frame.contentWindow.matchMedia("(min-width: 1px)").matches);
    }, "matchMedia should assume a 0x0 viewport in display: none iframes");
  }
</script>
<iframe style="display: none" onload="frameLoaded(this)"></iframe>
